+++ src/components/carousel/simple-carousel.js
@@ -37,6 +37,10 @@ var _typography = require("../../helpers/utils/typography");
 
 var _visuallyHidden = require("../../components/accessibility/visually-hidden");
 
+function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+
+function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
+
 var StyledContainer = _styledComponents.default.div.withConfig({
   displayName: "simple-carousel__StyledContainer",
   componentId: "t6k8ig-0"
@@ -64,16 +68,16 @@ var StyledPaginationButton = _styledComponents.default.button.withConfig({
   return activePaginationColor;
 });
 
-var SimpleCarousel =
-/*#__PURE__*/
-function (_Component) {
+var SimpleCarousel = /*#__PURE__*/function (_Component) {
   (0, _inherits2.default)(SimpleCarousel, _Component);
+
+  var _super = _createSuper(SimpleCarousel);
 
   function SimpleCarousel(props) {
     var _this;
 
     (0, _classCallCheck2.default)(this, SimpleCarousel);
-    _this = (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(SimpleCarousel).call(this, props));
+    _this = _super.call(this, props);
 
     _this.showPagination = function () {
       return _this.state.totalPagesCount > 1;