+++ src/components/form/text-input-with-unit-form/index.js
@@ -41,6 +41,10 @@ var _button = require("../../../components/buttons/button");
 
 var _screenConfig = require("../../../constants/screen-config");
 
+function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+
+function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
+
 var StyledFormContainer = _styledComponents.default.form.withConfig({
   displayName: "text-input-with-unit-form__StyledFormContainer",
   componentId: "wjoixz-0"
@@ -69,14 +73,14 @@ var StyledButton = (0, _styledComponents.default)(function (_ref3) {
   return version === 'tiny' && (0, _styledComponents.css)(["width:100%;"]);
 });
 
-var TextInputWithUnitForm =
-/*#__PURE__*/
-function (_PureComponent) {
+var TextInputWithUnitForm = /*#__PURE__*/function (_PureComponent) {
   (0, _inherits2.default)(TextInputWithUnitForm, _PureComponent);
+
+  var _super = _createSuper(TextInputWithUnitForm);
 
   function TextInputWithUnitForm() {
     (0, _classCallCheck2.default)(this, TextInputWithUnitForm);
-    return (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(TextInputWithUnitForm).apply(this, arguments));
+    return _super.apply(this, arguments);
   }
 
   (0, _createClass2.default)(TextInputWithUnitForm, [{