+++ src/components/heroes/hero.js
@@ -39,17 +39,21 @@ var _colorsConfig = _interopRequireDefault(require("../../constants/colors-confi
 
 var _right2, _content, _left3, _right3, _image;
 
+function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+
+function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
+
 var Grid = (0, _radium.default)(_grid.Grid);
 var GridCol = (0, _radium.default)(_grid.GridCol);
 
-var HeroBase =
-/*#__PURE__*/
-function (_Component) {
+var HeroBase = /*#__PURE__*/function (_Component) {
   (0, _inherits2.default)(HeroBase, _Component);
+
+  var _super = _createSuper(HeroBase);
 
   function HeroBase() {
     (0, _classCallCheck2.default)(this, HeroBase);
-    return (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(HeroBase).apply(this, arguments));
+    return _super.apply(this, arguments);
   }
 
   (0, _createClass2.default)(HeroBase, [{