+++ src/components/videos/video/index.js
@@ -19,7 +19,7 @@ var _createClass2 = _interopRequireDefault(require("@babel/runtime/helpers/creat
 
 var _possibleConstructorReturn2 = _interopRequireDefault(require("@babel/runtime/helpers/possibleConstructorReturn"));
 
-var _getPrototypeOf3 = _interopRequireDefault(require("@babel/runtime/helpers/getPrototypeOf"));
+var _getPrototypeOf2 = _interopRequireDefault(require("@babel/runtime/helpers/getPrototypeOf"));
 
 var _inherits2 = _interopRequireDefault(require("@babel/runtime/helpers/inherits"));
 
@@ -34,6 +34,10 @@ var _typography = require("../../../helpers/utils/typography");
 var _colorsConfig = _interopRequireDefault(require("../../../constants/colors-config"));
 
 var _reactElements = require("../../../helpers/react/react-elements");
+
+function _createSuper(Derived) { return function () { var Super = (0, _getPrototypeOf2.default)(Derived), result; if (_isNativeReflectConstruct()) { var NewTarget = (0, _getPrototypeOf2.default)(this).constructor; result = Reflect.construct(Super, arguments, NewTarget); } else { result = Super.apply(this, arguments); } return (0, _possibleConstructorReturn2.default)(this, result); }; }
+
+function _isNativeReflectConstruct() { if (typeof Reflect === "undefined" || !Reflect.construct) return false; if (Reflect.construct.sham) return false; if (typeof Proxy === "function") return true; try { Date.prototype.toString.call(Reflect.construct(Date, [], function () {})); return true; } catch (e) { return false; } }
 
 var StyledContainer = _styledComponents.default.div.withConfig({
   displayName: "video__StyledContainer",
@@ -67,14 +71,12 @@ var StyledVideo = _styledComponents.default.video.withConfig({
   return !controls && (0, _styledComponents.css)(["object-fit:cover;"]);
 });
 
-var Video =
-/*#__PURE__*/
-function (_PureComponent) {
+var Video = /*#__PURE__*/function (_PureComponent) {
   (0, _inherits2.default)(Video, _PureComponent);
 
-  function Video() {
-    var _getPrototypeOf2;
+  var _super = _createSuper(Video);
 
+  function Video() {
     var _this;
 
     (0, _classCallCheck2.default)(this, Video);
@@ -83,7 +85,7 @@ function (_PureComponent) {
       args[_key] = arguments[_key];
     }
 
-    _this = (0, _possibleConstructorReturn2.default)(this, (_getPrototypeOf2 = (0, _getPrototypeOf3.default)(Video)).call.apply(_getPrototypeOf2, [this].concat(args)));
+    _this = _super.call.apply(_super, [this].concat(args));
     _this.video = (0, _react.createRef)();
     _this.state = {
       showPlayer: false