+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `cssPropertyDistributor` utility.
+- Fix: Add API check to `ContinuousIntersectionObserver` utils.
 
 ## [2.60.3] - 2020-04-01