+++ CHANGELOG.md
@@ -4,8 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `cssPropertyDistributor` utility.
-- Fix: Add API check to `ContinuousIntersectionObserver` utils.
+## [2.61.0] - 2020-04-02
+
+Feature:
+- Add `cssPropertyDistributor` utility.
+
+Fix:
+- Add API check to `ContinuousIntersectionObserver` utils.
 
 ## [2.60.3] - 2020-04-01